+++ TinkoffASDKUI/TinkoffASDKUI/Modules/CardPayment/Presenter/CardPaymentPresenter.swift
@@ -83,6 +83,11 @@ extension CardPaymentPresenter {
         }
     }
 
+    func viewDidAppear() {
+        cardFieldPresenter.activate(textFieldType: .cardNumber)
+        savedCardPresenter?.activateCVCField()
+    }
+
     func closeButtonPressed() {
         router.closeScreen()
     }