+++ TinkoffASDKUI/TinkoffASDKUI/Modules/RecurrentPayment/Services/IRecurrentPaymentTableContentProvider.swift
@@ -1,0 +1,10 @@
+//
+//  IRecurrentPaymentTableContentProvider.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 18.04.2023.
+//
+
+import UIKit
+
+protocol IRecurrentPaymentTableContentProvider: ITableContentProvider where CellType == RecurrentPaymentCellType {}