+++ TinkoffASDKUI/TinkoffASDKUI/Modules/RecurrentPayment/View/IRecurrentPaymentViewInput.swift
@@ -6,11 +6,17 @@
 //
 
 protocol IRecurrentPaymentViewInput: AnyObject {
-    func showCommonSheet(state: CommonSheetState)
+    func showCommonSheet(state: CommonSheetState, animatePullableContainerUpdates: Bool)
     func hideCommonSheet()
 
     func hideKeyboard()
 
     func reloadData()
     func closeView()
+}
+
+extension IRecurrentPaymentViewInput {
+    func showCommonSheet(state: CommonSheetState) {
+        showCommonSheet(state: state, animatePullableContainerUpdates: true)
+    }
 }